Added inference configs
