+++ configs/paths/inference_paths.yaml
@@ -1,0 +1,1 @@
+final_model_path: final_model/model.pkl
