+++ configs/paths/train_paths.yaml
@@ -6,7 +6,6 @@ data_paths:
 
 model_paths:
   model_path: saved_model/model.pkl
-  final_model_path: final_model/model.pkl
 
 plot_paths:
   plot_dirname: plots
