+++ requirements.txt
@@ -1,4 +1,4 @@
-numpy==1.22.4
+numpy~=1.26.2
 Pillow~=8.4.0
 matplotlib~=3.5.1
 pynput~=1.7.6
@@ -6,7 +6,7 @@ opencv-python==4.7.0.72
 opencv-contrib-python
 omegaconf~=2.3.0
 PyYAML~=6.0
-torchvision~=0.13.0
-scipy~=1.7.3
+torchvision~=0.14.0+cu117
+scipy~=1.11.4
 hydra-core~=1.3.2
 torch~=1.13.0+cu117
