+++ src/__init__.py
@@ -1,0 +1,3 @@
+import os
+
+SRC_DIR_PATH = os.getcwd()
