+++ src/consts.py
@@ -1,2 +1,6 @@
+import os
+
+from src import SRC_DIR_PATH
+
 TRAIN_CONFIG_NAME = r'train.yaml'
-TRAIN_CONFIG_PATH = r'../../configs/'
+TRAIN_CONFIG_PATH = os.path.join(SRC_DIR_PATH, r'../../configs/')
